scan: rework obsolence calculation to handle same subject patches
Pull requests often result the same subject name.  Instead of always looking
for the newest version, maintain a list of dates and versions and look for
a newer version in that list.  IOW, if you have three patches of:

    20130101   [PULL] Block patches
    20130103   [PULL v2] Block patches
    20130110   [PULL] Block patches

Instead of marking both 20130101 and 20130110 obsolete, form a list of:

    [(20130101, 1), (20130103, 2), (20130110, 1)]

Then look for the current message in the list examining the next message to
see if it's obsolete or not.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>